+++ python/mocpy/tmoc/tmoc.py
@@ -70,29 +70,18 @@ class TimeMOC(AbstractMOC):
     # Default observation time : 30 min
     DEFAULT_OBSERVATION_TIME = TimeDelta(30 * 60, format="sec", scale="tdb")
 
-    __create_key = object()
-
-    def __init__(self, create_key, store_index):
+    def __init__(self, store_index):
         """Is a Time Coverage (T-MOC).
 
         Args:
-            create_key: Object ensure __init__ is called by super-class/class-methods only
             store_index: index of the S-MOC in the rust-side storage
         """
-        super().__init__(
-            AbstractMOC._create_key,
-            TimeMOC.__create_key,
-            store_index,
-        )
-        if create_key != TimeMOC.__create_key:
-            raise PermissionError(
-                "T-MOC instantiation is only allowed by class or super-class methods",
-            )
+        self.store_index = store_index
 
     @property
     def max_order(self):
         """Depth/order of the T-MOC."""
-        depth = mocpy.get_tmoc_depth(self._store_index)
+        depth = mocpy.get_tmoc_depth(self.store_index)
         return np.uint8(depth)
 
     @classmethod
@@ -124,12 +113,12 @@ class TimeMOC(AbstractMOC):
 
     def to_time_ranges(self):
         """Return the time ranges this TimeMOC contains."""
-        return microseconds_to_times(mocpy.to_ranges(self._store_index))
+        return microseconds_to_times(mocpy.to_ranges(self.store_index))
 
     @property
     def to_depth61_ranges(self):
         """Return the list of ranges this TimeMOC contains, in microsec since JD=0."""
-        return mocpy.to_ranges(self._store_index)
+        return mocpy.to_ranges(self.store_index)
 
     def degrade_to_order(self, new_order):
         """
@@ -148,8 +137,8 @@ class TimeMOC(AbstractMOC):
         moc : `~mocpy.tmoc.TimeMOC`
             The degraded MOC.
         """
-        index = mocpy.degrade(self._store_index, new_order)
-        return TimeMOC(TimeMOC.__create_key, index)
+        index = mocpy.degrade(self.store_index, new_order)
+        return TimeMOC(index)
 
     @classmethod
     def new_empty(cls, max_depth):
@@ -167,7 +156,7 @@ class TimeMOC(AbstractMOC):
             The MOC
         """
         index = mocpy.new_empty_tmoc(np.uint8(max_depth))
-        return cls(cls.__create_key, index)
+        return cls(index)
 
     @classmethod
     def from_depth61_ranges(cls, max_depth, ranges):
@@ -196,7 +185,7 @@ class TimeMOC(AbstractMOC):
             ranges = ranges.astype(np.uint64)
 
         index = mocpy.from_time_ranges_array2(np.uint8(max_depth), ranges)
-        return cls(cls.__create_key, index)
+        return cls(index)
 
     @classmethod
     def from_times(cls, times, delta_t=DEFAULT_OBSERVATION_TIME):
@@ -221,7 +210,7 @@ class TimeMOC(AbstractMOC):
 
         depth = TimeMOC.time_resolution_to_order(delta_t)
         store_index = mocpy.from_time_in_microsec_since_jd_origin(depth, times)
-        return cls(cls.__create_key, store_index)
+        return cls(store_index)
 
     @classmethod
     def from_time_ranges(cls, min_times, max_times, delta_t=DEFAULT_OBSERVATION_TIME):
@@ -262,7 +251,7 @@ class TimeMOC(AbstractMOC):
             min_times,
             max_times,
         )
-        return cls(cls.__create_key, store_index)
+        return cls(store_index)
 
     @classmethod
     def from_time_ranges_approx(
@@ -305,7 +294,7 @@ class TimeMOC(AbstractMOC):
             )
 
         store_index = mocpy.from_time_ranges(depth, min_times, max_times)
-        return cls(cls.__create_key, store_index)
+        return cls(store_index)
 
     @classmethod
     def from_stmoc_space_fold(cls, smoc, stmoc):
@@ -317,8 +306,8 @@ class TimeMOC(AbstractMOC):
         smoc : `~mocpy.moc.Moc`
         stmoc : `~mocpy.stmoc.STMoc`
         """
-        store_index = mocpy.project_on_first_dim(smoc._store_index, stmoc._store_index)
-        return cls(cls.__create_key, store_index)
+        store_index = mocpy.project_on_first_dim(smoc.store_index, stmoc.store_index)
+        return cls(store_index)
 
     def _process_degradation(self, another_moc, order_op):
         """
@@ -452,7 +441,7 @@ class TimeMOC(AbstractMOC):
 
         """
         return TimeDelta(
-            mocpy.ranges_sum(self._store_index) / 1e6,
+            mocpy.ranges_sum(self.store_index) / 1e6,
             format="sec",
             scale="tdb",
         )
@@ -521,7 +510,7 @@ class TimeMOC(AbstractMOC):
         # of the TimeMoc object
         pix_arr = times_to_microseconds(times)
 
-        mask = mocpy.filter_time(self._store_index, pix_arr)
+        mask = mocpy.filter_time(self.store_index, pix_arr)
 
         if keep_inside:
             return mask
@@ -716,13 +705,13 @@ class TimeMOC(AbstractMOC):
         path = str(path)
         if format == "fits":
             index = mocpy.time_moc_from_fits_file(path)
-            return cls(cls.__create_key, index)
+            return cls(index)
         if format == "ascii":
             index = mocpy.time_moc_from_ascii_file(path)
-            return cls(cls.__create_key, index)
+            return cls(index)
         if format == "json":
             index = mocpy.time_moc_from_json_file(path)
-            return cls(cls.__create_key, index)
+            return cls(index)
         formats = ("fits", "ascii", "json")
         raise ValueError("format should be one of %s" % (str(formats)))
 
@@ -730,7 +719,7 @@ class TimeMOC(AbstractMOC):
     def _from_fits_raw_bytes(cls, raw_bytes):
         """Load MOC from raw bytes of a FITS file."""
         index = mocpy.time_moc_from_fits_raw_bytes(raw_bytes)
-        return cls(cls.__create_key, index)
+        return cls(index)
 
     @classmethod
     def from_string(cls, value, format="ascii"):  # noqa: A002
@@ -750,9 +739,9 @@ class TimeMOC(AbstractMOC):
         """
         if format == "ascii":
             index = mocpy.time_moc_from_ascii_str(value)
-            return cls(cls.__create_key, index)
+            return cls(index)
         if format == "json":
             index = mocpy.time_moc_from_json_str(value)
-            return cls(cls.__create_key, index)
+            return cls(index)
         formats = ("ascii", "json")
         raise ValueError("format should be one of %s" % (str(formats)))