+++ CHANGELOG.md
@@ -24,7 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * all methods of `MOC` with signatures like `function(self, lon, lat, **kwargs)` now accept both lists of coordinates and single coordinates
-* * `mocpy.stmoc.STMOC.from_spatial_coverages` also accepts single moc objects (had to be a list before)
+* `mocpy.stmoc.STMOC.from_spatial_coverages` also accepts single moc objects (had to be a list before)
+* `AbstractMOC` derives from metaclass `ABCMeta`
 
 ### Added